\name{gpuDeviceInfo}
\alias{gpuDeviceInfo}
\title{Retreiving GPU Device Information}

\description{
This function prints some information about the CUDA device used in the gpusim package. Information include memory limits, clock frequency, cuda version and number of SIMD cores.
}

\usage{
	gpuDeviceInfo() 
}


\value{Internal ID for given covariance model}

\author{Marius Appel \email{marius.appel@uni-muenster.de}}

\examples{
gpuDeviceInfo()
}

\keyword{cuda}
\keyword{gpu}
